pyproject.toml
[project]
name = "automated_capability_evaluation"
description = "A repository for research project on generating LLM evaluation benchmarks using LLMs"
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Omkar Dige", email = "omkar.dige@vectorinstitute.ai"},
    {name = "Arash Afkanpour", email = "arash.afkanpour@vectorinstitute.ai"},
]
dynamic = ["version"]
dependencies = [
    "datasets>=3.2.0",
    "google-cloud-storage>=3.0.0",
    "hydra-core>=1.3.2",
    "inspect-ai>=0.3.159",
    "langchain_openai>=0.3.6",
    "langchain>=0.3.19",
    "matplotlib>=3.10.0",
    "omegaconf>=2.3.0",
    "openai>=1.102.0",
    "ag2>=0.3.2",
    "autogen-ext[openai,anthropic]>=0.7.4",
    "anthropic>=0.64.0",
    "ratelimit>=2.2.1",
    "torchvision (>=0.21.0,<0.22.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "gpytorch (>=1.14,<2.0)",
    "ruff (>=0.11.4,<0.12.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "vec-inf (==0.6.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)",
    "google-genai (>=1.13.0,<2.0.0)",
    "langchain-anthropic (==0.3.12)",
    "pytest (==8.3.3)",
    "pytest-env (>=1.1.5,<2.0.0)",
    "langfuse>=2.0.0",
    "openlit>=1.35.0",
]

[project.urls]
repository = "https://github.com/VectorInstitute/automated_capability_evaluation"

[tool.poetry]
package-mode = false
version = "0.1"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pre-commit = "^4.0.0"
pytest-cov = "^3.0.0"
pytest-env = "^1.1.5"
pytest-mock = "^3.10.0"
codecov = "^2.1.13"
mypy = "^1.15.0"
ruff = ">=0.11.4,<0.12.0"
nbqa = { version = "^1.7.0", extras = ["toolchain"] }
pip-audit = "^2.7.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
numpydoc = "^1.2"
sphinx = "^7.2.5"
sphinxcontrib-apidoc = "^0.4.0"
sphinx-autodoc-typehints = "^1.24.0"
myst-parser = "^2.0.0"
sphinx-design = "^0.5.0"
sphinx-copybutton = "^0.5.0"
sphinx-autoapi = "^2.0.0"
nbsphinx = "^0.9.3"
ipython = "^8.8.0"
ipykernel = "^6.23.0"
furo = "^2024.01.29"

[tool.mypy]
ignore_missing_imports = true
install_types = true
pretty = true
namespace_packages = true
explicit_package_bases = true
non_interactive = true
warn_unused_configs = true
allow_any_generics = false
allow_subclassing_any = false
allow_untyped_calls = false
allow_untyped_defs = false
allow_incomplete_defs = false
check_untyped_defs = true
allow_untyped_decorators = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true
extra_checks = true

[tool.ruff]
include = ["*.py", "pyproject.toml", "*.ipynb"]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "A", # flake8-builtins
    "B", # flake8-bugbear
    "COM", # flake8-commas
    "C4", # flake8-comprehensions
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "ICN", # flake8-import-conventions
    "Q", # flake8-quotes
    "RSE", # flake8-raise
    "D", # pydocstyle
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "W", # pycodestyle
    "N", # pep8-naming
    "ERA", # eradicate
    "PL", # pylint
]
fixable = ["A", "B", "COM", "C4", "RET", "SIM", "ICN", "Q", "RSE", "D", "E", "F", "I", "W", "N", "ERA", "PL"]
ignore = [
    "B905", # `zip()` without an explicit `strict=` parameter
    "E501", # line too long
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
    "PLR2004", # Replace magic number with named constant
    "PLR0913", # Too many arguments
    "COM812", # Missing trailing comma
    "PLR0915", # Too many statements
    "PLR0912", # Too many branches
]

# Ignore import violations in all `__init__.py` files.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]
"tests/src/seed_capabilities/math/competition_math_algebra/capability.py" = ["D100", "D101", "D102"]
"tests/src/capabilities_t2/math/competition_math_combinatorial_challenges/capability.py" = ["D100", "D101", "D102"]
"src/run.py" = ["ERA001"]
"src/lbo.py" = ["ERA001"]
"src/utils/capability_utils.py" = ["ERA001"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["X*", "setUp"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pycodestyle]
max-doc-length = 88

[tool.pytest.ini_options]
markers = [
    "integration_test: marks tests as integration tests",
]
env = [
    "VEC_INF_LOG_DIR=DUMMY",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage]
    [tool.coverage.run]
    source=["src"]
    omit=["tests/*", "*__init__.py"]

    [tool.coverage.report]
    # Ignore errors such as "No source for code" that can happen when runtime-generated
    # files are deleted before the coverage report phase. This prevents coverage from
    # exiting with status 1 even though the test suite passed.
    ignore_errors = true

[build-system]
requires = ["poetry-core>=2.0.1"]
build-backend = "poetry.core.masonry.api"
